feat!(fpss): rename the protocol parse-error event to ParseError
The FPSS parse-error event class is named ParseError in every binding (kind tag parse_error), so no binding ships a class answering to the bare Error name: the Python exception hierarchy roots unambiguously at ThetaDataError, the JS global Error is no longer shadowed by an SDK interface, and C / C++ expose TdxFpssParseError / tdx::FpssParseError. The TdxFpssEventKind C discriminants renumber with the sorted variant set.

The event-schema generators gain an explicit mapping between schema event names and the core FpssControl variants, so a public event class can be named independently of the Rust enum variant that backs it. The Python stub drops the FpssParseError alias and its stubtest allowlist entry; the stub now matches the runtime symbol for symbol.

+++ crates/thetadatadx/build_support_bin/fpss_events/ffi_c.rs
@@ -40,8 +40,9 @@ fn render_contract_struct_c() -> &'static str {
  * NUL-terminated C string (may be null when not yet resolved);\n\
  * optional option fields use a tagged-present bool because C has no\n\
  * Option<T>. `right` is `'C'` / `'P'` (NUL when `has_right` is false)\n\
- * — the same notation the public option builder takes. Layout is\n\
- * byte-identical to Rust's #[repr(C)] TdxContract.\n\
+ * and `strike` is the option strike in dollars — the same notation\n\
+ * the public option builder takes. Layout is byte-identical to Rust's\n\
+ * #[repr(C)] TdxContract.\n\
  */\n\
 typedef struct {\n\
     const char *symbol;\n\
@@ -51,7 +52,7 @@ typedef struct {\n\
     bool has_right;\n\
     char right;\n\
     bool has_strike;\n\
-    int32_t strike;\n\
+    double strike;\n\
 } TdxContract;\n\n"
 }
